mypo-34 refactoring rebalancer
