+++ mypo/runner.py
@@ -6,9 +6,10 @@ import numpy as np
 import numpy.typing as npt
 import pandas as pd
 
+from mypo.rebalancer.rebalancer import Rebalancer
+
 from .common import calc_capital_gain_tax, calc_fee, calc_income_gain_tax, safe_cast
 from .market import Market
-from .rebalancer import Rebalancer
 from .reporter import Reporter
 from .settings import Settings
